backend/requirements.txt
flask
requests
json5
jsonschema
flask-cors
flask_security
flask_mongoengine
passlib
pyyaml
pymongo
Pillow
whoosh
